+++ main.py
@@ -20,7 +20,6 @@ from discord.ext import commands, tasks
 from PIL import Image, ImageDraw, ImageFont
 import os
 import requests
-import json
 from datetime import datetime, timezone
 import time
 import sqlite3
@@ -28,7 +27,8 @@ import sqlite3
 from secret_const import TOKEN
 
 from const import CACHE_DIR_PFP, COLORS, LEADERBOARD_PIC, DEFUALT_PROFILE_PIC, LOG_CHANNEL_ID, ADMIN_LOG_CHANNEL_ID, CENSORSHIP_CHANNEL_ID, CARD_DATA_JSON_PATH, CARD_DATA_IMAGES_PATH, DATABASE_PATH
-from helperFunctions import xpToLevel, updateXpAndCheckLevelUp, copyCard, checkDatabase, censorMessage, checkLoginRemindersAndSend
+
+from helperFunctions.main import xpToLevel, updateXpAndCheckLevelUp, copyCard, checkDatabase, censorMessage, checkLoginRemindersAndSend
 
 from adminCommands.set import set
 from adminCommands.stats import stats
@@ -49,8 +49,6 @@ from commands.credits import credits
 
 from floor10_game_concept import guess_the_number_command
 
-from generateCardAI import genAiCard
-from cardImageMaker import makeCardFromJson
 from fight import ChallengeView
 
 class MyBot(commands.Bot):
@@ -517,49 +515,5 @@ async def challenge(interaction: discord.Interaction, member: discord.Member):
         challenge_view = ChallengeView(challenger=interaction.user, challenged=member, timeout_message=challenge_message)
         challenge_view.timeout_message = challenge_message
         await challenge_message.edit(view=challenge_view)
-
-@bot.tree.command(name="generatecard", description="Generate a custom playing card. (demo)")
-@app_commands.describe(prompt="Choose the type for the card.")
-async def genCard(interaction: discord.Interaction, prompt: str = "prompt"):
-    
-    # Defer the response to allow more time for processing
-    await interaction.response.defer()
-    output = await genAiCard(prompt) 
-    # returns a list, the first item is the json data, the second is the image url
-    card = await makeCardFromJson(output[0], output[1]) 
-    # output[1] is doing nothing we over write the varible in the file
-
-    # Update the user's items in the database
-    with sqlite3.connect(DATABASE_PATH) as conn:
-        cursor = conn.cursor()
-
-        # Fetch the current highest itemID
-        cursor.execute("SELECT MAX(itemId) FROM cards")
-        result = cursor.fetchone()
-        currentItemId = result[0] + 1 if result[0] is not None else 1
-        cursor.execute(
-            "INSERT INTO cards (itemName, userId, cardId) VALUES (?, ?, ?)", 
-            (output[0]['name'], interaction.user.id, currentItemId))
-        conn.commit()
-        # Save the card data as a JSON file
-        output_filename = f"{currentItemId}.json"
-        output_path = os.path.join(CARD_DATA_JSON_PATH, output_filename)
-        if not os.path.exists(CARD_DATA_JSON_PATH):
-            os.makedirs(CARD_DATA_JSON_PATH)
-        with open(output_path, 'w') as json_file:
-            json.dump(output[0], json_file, indent=4)
-        # Save the card image
-        card_name = f"{currentItemId}.png"
-        
-        if not os.path.exists(CARD_DATA_IMAGES_PATH):
-            os.makedirs(CARD_DATA_IMAGES_PATH)
-        cardFilePath = f'{CARD_DATA_IMAGES_PATH}/{card_name}'
-        
-        card.save(cardFilePath)
-
-    file = discord.File(cardFilePath, filename="card.png")
-    # Edit the initial deferred response to include the embed with the image file
-    await interaction.followup.send(file=file)
-
 
 bot.run(TOKEN)
